requirements.txt
playwright>=1.49.1
beautifulsoup4>=4.12.3
tqdm>=4.66.2
python-dotenv>=1.0.1
requests>=2.31.0
aiohttp>=3.10.11
telethon>=1.38.1
playwright-stealth>=1.0.6
aiogram>=3.4.1
psutil>=5.9.8
